Private types

STLCutters.AbstractPlane

STLCutters.CartesianPlane

STLCutters.Face

STLCutters.Plane

STLCutters.STL

STLCutters.STLTopology

STLCutters.Segment

STLCutters.SubtriangulationLabels

STLCutters.Tetrahedron

STLCutters.Triangle

STLCutters.AbstractPlane — Type

abstract type AbstractPlane

Plane type used to compute intersections. It can be accessed by normal and origin.

STLCutters.CartesianPlane — Type

struct CartesianPlane{D,T}<:AbstractPlane{D,T}

Plane defined by a Cartesian axis and one single Cartesian coordinate.

STLCutters.Face — Type

abstract type Face{Df,Dp}

Face type used to compute intersections with multiple dispatching.

STLCutters.Plane — Type

struct Plane{D,T}<:AbstractPlane{D,T}

Plane defined by an origin point and a normal vector.

STLCutters.STL — Type

const STL = STLTopology

Alias STLTopology as STL.

STLCutters.STLTopology — Type

struct STLTopology{Dc,Dp,T} <: GridTopology{Dc,D}

STLTopology is a Gridap.GridTopology without lazy arrays. It reduces computations when accessing their properties. It is used to save the STL geometries. It is alliased as STL.

STLCutters.Segment — Type

struct Segment{D,T}<:Face{1,D}

1-dimensional Face defined by two points.

STLCutters.SubtriangulationLabels — Type

SubtriangulationLabels

Labels for the sub grids of subtriangulate.

It stores the following fields:

cell_to_bgcell: subcell to background cell

cell_to_io: subcell to IN/OUT

face_to_stlface: subface to STL face

face_to_bgcell: subface to background cell

face_to_ios: subface to surface source (IN,OUT,SKIN)

bgcell_to_ioc: background cell to IN/OUT/CUT

bgface_to_ioc: background face to IN/OUT/CUT

bface_to_lbgface: subface to local background face

bface_to_bgcell: subface to background cell

bface_to_io: subface to IN/OUT

STLCutters.Tetrahedron — Type

struct Tetrahedron{D,T}<:Face{3,D}

3-dimensional simplex Face defined by four points.

STLCutters.Triangle — Type

struct Triangle{D,T}<:Face{2,D}

2-dimensional simplex Face defined by three points.